+++ lib/laflem/collections/python/const.py
@@ -1,0 +1,10 @@
+"""
+Constants for the python collection.
+"""
+
+PACKAGE_NAME = "python"
+
+PYTHON_CACHE_LOCATION = {
+    "linux": "~/.cache/pip",
+    "darwin": "~/Library/Caches/pip",
+}